backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
greenlet
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
httpx
alembic
psycopg2-binary
aiosqlite
supabase
